hotfix(test): Remove obsolete test scenario
After updating the time-series extraction mechanism, the function 'get_agent_time_series_data()' now returns two additional columns: AgentID and Step. This is important for further analysis, so that we can differentiate which rows come from which agent/step. For now, the test scenario becomes obsolete. In the future, I may consider revising the approach and reactivate the test.